@@ -1,7 +1,7 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Component } from '@angular/core';
-import { MatDialog } from '@angular/material/dialog';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
 import { AddToListDialogComponent } from '../add-to-list-dialog/add-to-list-dialog.component';
 import { CartService } from '../cart.service';
 import { ComicService } from '../comic.service';
